requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.1
Django==3.0.5
django-appconf==1.0.4
django-braces==1.14.0
django-cleanup==5.1.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-embed-video==1.3.3
django-haystack==3.0
django-model-utils==4.0.0
django-paypal==1.0.0
django-ratings==0.3.7
django-sphinx==2.2.4
django-video-encoding==0.4.0
filelock==3.0.12
idna==2.10
importlib-metadata==3.4.0
importlib-resources==5.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paystackapi==2.0.0
Pillow==5.4.1
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
requests==2.24.0
shutilwhich==1.1.0
six==1.15.0
South==1.0.2
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.55.0
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.4.0
Whoosh==2.7.4
zipp==3.4.0